server/plugins.c: Fix pointer type mismatch while processing    CFPython.RegisterCommand.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/trunk/crossfire@3090 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ server/plugins.c
@@ -78,7 +78,7 @@ CommArray_s *find_plugin_command(char *cmd, object *op)
         {
             RTNCmd.name = (char *)(RTNValue->Value[0]);
             RTNCmd.func = (CommFunc)(RTNValue->Value[1]);
-            RTNCmd.time = *(float *)(RTNValue->Value[2]);
+            RTNCmd.time = *(double *)(RTNValue->Value[2]);
             printf("RTNCMD: name %s, time %f\n", RTNCmd.name, RTNCmd.time);
             return &RTNCmd;
         };
